feat: vmind数据模块开发&划分gpt和skylark
