+++ packages/vmind/package.json
@@ -71,6 +71,7 @@
     "axios": "^1.4.0",
     "exceljs": "^4.3.0",
     "lodash": "4.17.21",
-    "json5": "~2.2.3"
+    "json5": "~2.2.3",
+    "dayjs": "~1.11.10"
   }
 }
